Ammo system rework V1
Summary:

The player can grab up to two ammo boxes, and grabbing the third one drops the currently selected ammo box. Pressing Q swaps selected ammo boxes

The player can reload the gun one bullet at a time, up to 6 bullets in its cylinder. While the currently selected chamber is loaded, reloading rotates the cylinder by 1 chamber clockwise. Shooting the gun causes the cylinder to rotate by 1 chamber clockwise, firing a bullet if the selected chamber is loaded. Finally, the cylinder can be rotated in both clockwise and counter-clockwise directions with the mouse wheel.

Finally, the color of the cylinder and the gun changes according to the color of the loaded bullet in the selected chamber.

Scripts Edited:

- RevolverChamber
- HUDController
- AmmoManager

Prefabs:

- Added Yellow Bullet
- Updated various components of RevolverPlayer
- Added Rigidbody to Ammo Boxes of all colors
- Changed Collider size of all Ammo Boxes