Sound Effect for empty chamber added
When player fires with no bullet in the gun, empty revolver audio clip is played.
User Story: As a player, I want to hear an "empty clip" sound effect when I try to shoot with an empty chamber, so that I get auditory feedback indicating that my gun is out of bullets.