Animated Models Idle and Running
Added the animation for the models from blender into unity and add components so they work in Untiy
User Story: As an Artist, I want to have an Idle animation, So that the Enemy looks alive and has behaviour.
User Story: As an Artist, I want to have a Run animation, So that the enemy can move around the map.
(7 hours)