Added AmmoBox Pickup and Switch sounds
Added AmmoBox Pickup and Switch sounds. AmmoSwitch sound does not play when when there are no ammo boxes in the inventory.
User Story: As a player, I want to hear a sound effect when switch ammo boxs so that I receive clear audio feedback, enhancing the game's immersive experience.
User Story: As a player, I want to hear a sound effect when I pick up an ammo box so that I receive clear audio feedback, enhancing the game's immersive experience.