Laser Wall Sound added
Sound for the laser wall sound added.

User Story:As a player, I want to hear a distinct sound effect when I encounter the laser wall, so that I am alerted to its presence and can associate the sound with the laser's behavior or danger.
Estd Time: 3 Hours
Actual TIme: 9 Hours